requirements.txt
torch
transformers
accelerate
huggingface_hub
bitsandbytes
colorama
python-dotenv
fastapi
uvicorn[standard]
pydantic
httpx
